Working of fixing memory leaks and norm errors
