+++ minishell.c
@@ -6,7 +6,7 @@
 /*   By: paulorod <paulorod@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/07/20 15:31:24 by paulorod          #+#    #+#             */
-/*   Updated: 2023/09/01 16:29:45 by paulorod         ###   ########.fr       */
+/*   Updated: 2023/09/04 14:45:31 by paulorod         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -56,12 +56,19 @@ t_cmd	*command_parser(char *cmd_line, t_shell *shell)
 {
 	t_cmd	*cmd_struct;
 	char	**tokens;
+	int		i;
 
-	cmd_struct = ft_calloc(sizeof(t_cmd), 1);
 	tokens = create_cmd_tokens(cmd_line, shell);
 	cmd_struct = create_cmd_list(tokens, shell);
 	cmd_struct->fd[1] = 1;
-	//free(cmd_line);
+	free(cmd_line);
+	i = 0;
+	while (tokens[i])
+	{
+		free(tokens[i]);
+		i++;
+	}
+	free(tokens);
 	return (cmd_struct);
 }
 
@@ -85,7 +92,7 @@ void	shell_loop(t_shell *shell)
 			if (ft_strlen(command) > 0)
 				shell->cmd = command_parser(command, shell);
 			handle_commands(shell);
-			//free_cmd(shell->cmd);
+			free_cmd(shell->cmd);
 		}
 	}
 }
