+++ srcs/commands/commands.c
@@ -6,7 +6,7 @@
 /*   By: paulorod <paulorod@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/08/10 16:16:56 by paulorod          #+#    #+#             */
-/*   Updated: 2023/09/04 12:31:54 by paulorod         ###   ########.fr       */
+/*   Updated: 2023/09/04 15:25:58 by paulorod         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -38,53 +38,6 @@ int	create_command_process(t_cmd *cmd, char **env)
 	}
 	return (WEXITSTATUS(status));
 }
-
-/*Replace environment variables with their value*/
-char	*handle_envs(char *token, t_shell *shell)
-{
-	char	*env_value;
-
-	env_value = extend_env_vars(token, shell, false);
-	if (env_value && *env_value)
-	{
-		free(token);
-		return (env_value);
-	}
-	else if (env_value)
-		free(env_value);
-	return ("");
-}
-
-/* OLD CODE
-Create command struct
-char	**old_create_cmd(char *_cmd, t_shell *shell)
-{
-	char			**cmd;
-	int				pos;
-	size_t			i;
-	unsigned int	j;
-
-	cmd = alloc_cmd(_cmd);
-	i = 0;
-	j = 0;
-	pos = 0;
-	while (_cmd[i])
-	{
-		if (_cmd[i] == ' ')
-		{
-			cmd[pos] = handle_envs(ft_substr(_cmd, j, i - j), shell);
-			pos++;
-			j = i + 1;
-		}
-		else if (!_cmd[i + 1])
-		{
-			cmd[pos] = handle_envs(ft_substr(_cmd, j, i + 1 - j), shell);
-			j = i + 1;
-		}
-		i++;
-	}
-	return (cmd);
-}*/
 
 /*Get command type indentifier*/
 enum e_identifiers	get_cmd_type(char *token)
@@ -118,6 +71,25 @@ t_cmd	*create_token_cmd(char *token)
 	return (command);
 }
 
+/*Get alloc size for command array*/
+int	get_cmd_size(char **tokens)
+{
+	int	i;
+	int	size;
+
+	i = 0;
+	size = 0;
+	while (tokens[i])
+	{
+		if (!is_special_char(tokens[i], 0, NULL))
+			size++;
+		else
+			break ;
+		i++;
+	}
+	return (size + 1);
+}
+
 /*Create linked list of command structs*/
 t_cmd	*create_cmd_list(char **tokens, t_shell *shell)
 {
@@ -128,7 +100,7 @@ t_cmd	*create_cmd_list(char **tokens, t_shell *shell)
 	i = 0;
 	j = 0;
 	command = ft_calloc(sizeof(t_cmd), 1);
-	command->cmd = ft_calloc(sizeof(char *), 1);
+	command->cmd = ft_calloc(sizeof(char *), get_cmd_size(tokens));
 	while (tokens[i])
 	{
 		if (!is_special_char(tokens[i], 0, NULL))
@@ -142,7 +114,8 @@ t_cmd	*create_cmd_list(char **tokens, t_shell *shell)
 			command->next = create_token_cmd(tokens[i]);
 			command->next->next = create_cmd_list(&tokens[i + 1], shell);
 			command = command->next;
-			command->cmd = ft_calloc(sizeof(char *), 1);
+			command->cmd = ft_calloc(sizeof(char *),
+					get_cmd_size(&tokens[i + 1]));
 		}
 		i++;
 	}
