+++ srcs/utils/utils.c
@@ -6,7 +6,7 @@
 /*   By: paulorod <paulorod@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/07/22 19:59:38 by vpacheco          #+#    #+#             */
-/*   Updated: 2023/08/30 13:45:08 by paulorod         ###   ########.fr       */
+/*   Updated: 2023/09/04 13:38:02 by paulorod         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -67,22 +67,25 @@ char	**fill_envs(const char **env)
 /*Free command struct*/
 void	free_cmd(t_cmd *cmd)
 {
-	int	i;
+	int		i;
+	t_cmd	*temp;
 
-	i = 0;
-	while (cmd->cmd[i])
+	while (cmd)
 	{
-		if (*(cmd->cmd[i]))
+		i = 0;
+		while (cmd->cmd[i])
 		{
-			//free(cmd->cmd[i]);
+			free(cmd->cmd[i]);
 			cmd->cmd[i] = NULL;
+			i++;
 		}
-		i++;
+		if (cmd->path)
+			free(cmd->path);
+		free(cmd->cmd);
+		temp = cmd->next;
+		free(cmd);
+		cmd = temp;
 	}
-	if (cmd->path)
-		free(cmd->path);
-	free(cmd->cmd);
-	free(cmd);
 }
 
 /*Custom getenv fucntion that searches our environment variable list*/
